hima1379 custom fork: SM-6/SPY-1/SPY-6/FCS Console additions (experimental)
Added features:
- SM-6 missile defense system (experimental, accuracy under improvement)
- SPY-1 / SPY-6 radar systems
- FCS Console multi-target tracking

3D model credits:
- SM-6 missile model: burt_cocaine (PlanetMinecraft, partially modified)
- SPG62 model: WTigerTw (Sketchfab)
- New fusion reactor model: HbmMods (GitHub)

Fork chain: hima1379 -> Alcater -> Drillgon200 -> HBMTheBobcat
License: GPL-3.0 + LGPL-3.0

+++ src/main/java/com/hbm/explosion/ExplosionFleija.java
@@ -2,6 +2,7 @@ package com.hbm.explosion;
 
 import com.hbm.blocks.generic.DecoBlockAlt;
 
+import net.minecraft.init.Blocks;
 import net.minecraft.nbt.NBTTagCompound;
 import net.minecraft.util.math.BlockPos;
 import net.minecraft.util.math.BlockPos.MutableBlockPos;
@@ -84,7 +85,7 @@ public class ExplosionFleija
 		if(shell == 0)
 			shell = 1;
 		int shell2 = this.shell * 2;
-		this.leg = (int) Math.floor((double) (this.n - (shell2 - 1) * (shell2 - 1)) / shell2);
+		this.leg = (int) Math.floor((this.n - (shell2 - 1) * (shell2 - 1)) / shell2);
 		this.element = (this.n - (shell2 - 1) * (shell2 - 1)) - shell2 * this.leg - this.shell + 1;
 		this.lastposX = this.leg == 0 ? this.shell : this.leg == 1 ? -this.element : this.leg == 2 ? -this.shell : this.element;
 		this.lastposZ = this.leg == 0 ? this.element : this.leg == 1 ? this.shell : this.leg == 2 ? -this.element : -this.shell;
